+++ src/coordinates/right_ascension.rs
@@ -1,13 +1,13 @@
 use crate::Float;
 
 pub struct RightAscension {
-    pub(super) hours: Float,
-    pub(super) minutes: Float,
-    pub(super) seconds: Float,
+    pub(super) hours: i8,
+    pub(super) minutes: i8,
+    pub(super) seconds: i8,
 }
 
 impl RightAscension {
-    pub const fn new(hours: Float, minutes: Float, seconds: Float) -> Self {
+    pub const fn new(hours: i8, minutes: i8, seconds: i8) -> Self {
         Self {
             hours,
             minutes,
